Change the `Implicit` and `CollisionDetector` traits.
Now the transformation matrix is not longer attached to the geometry.
Instead it is passed explictly as an argument to the support mapping function and to the collision
detectors.
This makes it easier to manipulate, and removes an unsafe hack on the one shot contact manifold
generator.

The geometries: `Transformed` and `Translated` are not longer needed. Removed.