Improve the one-shot contact manifold generation.
One-shot contact manifold generation rotates slightly one of the objects to find more contact
points.

Before: the object was rotated wrt its center of mass. This is not useful for long thin object.

After: the object is now rotated wrt the known contact point. The main benefit of this is that this
makes the known contact point invariant: we have the guarantee it will not be removed by the
incremental contact manifold during the one-shot generation.