Start the addition of bezier surfaces and curves.
Collision detection is not yet implemented for anything but `BezierSurface` vs. `Ball`.
Ray-casting is not implemented (it always returns `None` for now).

This also adds a new utility (needed for bezier curve/surface subdivision): `VecSlice/VecSliceMut`.
It is a wapper on `Vec` that give access on regularly-spaced elements using natural, continuous
indexing.