Add a procedural generation module.
This is a merge of the `nprocgen` project into `ncollide`.

The `nprocgen` project (and its
[repository](http://github.com/sebcrozet/nprocgen.git) will be
deleted soon.

This adds procedural generation functions into the `procedural`
module. At the moment it does not interact with other modules of
the project, but it will soon (it will define a trait that
`ncollide` geometries will implement to generate a polyhedric
representation of itself).

This also adds a new `utils` module contains miscellaneous
geometric functions that are useful here an there.