Add a trait implemented by geometries that can have a polyhedral representation.
This is the `ToTriMesh` trait.
It has one method that outputs a `TriMesh`.
The trait is not yet properly implemented for most meshes with a non-zero margin.

For unification purpoises, every texture coordinates on ncollide are 2d now (`Vec2` instead of `Vec3`).
If this is problematic for some user, everytning may be changed to Vec3 instead in the future.