Convex hull 3d: is no horizon is found, just stop the algorithm.
Not finding an horizon is a sign of a very bad problem, and most of the time it will crash with a
"affinely dependent facet" error. For the moment, we will just abort the algoritm at this point,
and return the current result.

A proper fix will have to be found though.