Fix for Issue #57
The problem described in issue #57 occurs because the object is removed on the remove() call then is not present during the update() call. This is because the narrow phase attempts to resolve colliding objects through the use of the ContactSignal. However, because the object was previously removed, one object in the pair in the callback handler is essentially null. The solution I made was to store a list of objects that are to be removed, but not actually do the removing until the end of update().