Handle object removal properly + fix typo: "defered -> deferred".
Objects removal was not handled properly at several places:
  − some collision pairs involving deleted objects were not properly removed by the broad phase.
  − the collision world removed the collision objects too soon (before the broad phase had a chance
  de report the proximity loss event).

The resolution here is to defer object removal from the collision world until an update is made.
Therefore, the `.remove(...)` method of `CollisionWorld` was renamed to `.deferred_remove(...)`. A
new method `.perform_removals_and_broad_phase()` actually applies the removal but, as a side
effect, triggers a broad phase update as well.

This fix also rename all the methods including the word `defered` to `deferred` (with two 'r').

Actually fix #57.