Add additional information to `PointProjection`
This makes it possibles for any code that does a point projection, to
add additional informatation. Since the type of this information is
dependent on the details of the projection, the feld in question is
generic. The type parameter that is needed for that is set to `()` by
default, to minimize the impact of this change.