Add trait `AdvancedPointQuery`
The trait's documentation describes why this is necessary, and how it
can be used.

Please note while it would be great to have a blanket implementation of
`PointQuery` for any type implementing `AdvancedTypeQuery`, adding such
a blanket implementation causes errors about conflicting implementations
for every implementation of `PointQuery`. I'm not sure why, as this
seems perfectly legitimate, and not conflicting at all.

For the moment I've chosen to forgo this convenience and just require
types that want to implement `AdvancedPointQuery` to implement
`PointQuery` in addition.