make collision world impl Sync Send
It changes some trait to require impl Sync and Send

I don't think if this is too restrictive for some usage but maybe.

dbvt was the only structure that need to be changed to impl Send and Sync

it now uses a DBVTLeaf struct that contains the
rc<refcell<DBVTLeafInner>>

to access the previously public fields it contains methods

those methods require a vec of dbvt that is needed for compile time
check.