Change query::visitors::*Collector to visitors
Instead of allocating a Vec to put the query results in, the Collectors
are changed to call a visitor instead. If the call-site does want a Vec,
they can easily collect themselves.
We ourselves do not actually need this, so we can remove a couple of
Vecs and corresponding allocations for some operations.

The *Collectors are also appropriately renamed to *Visitor.